Turn on -Wall and fix warnings.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7506 640d5228-2204-0410-8d71-8f46fa6850b4

+++ vardb/lib/category.c
@@ -169,7 +169,7 @@ char **VarDB_GetVariablesInCategory(int catNum)
   p = (char **)malloc(sizeof(char *));
 
   for (i = 0; i < VarDB_nRecords; ++i)
-    if (ntohl(((struct var_v2 *)VarDB)[i].Category) == catNum)
+    if (ntohl(((struct var_v2 *)VarDB)[i].Category) == (unsigned int)catNum)
       {
       p = (char **)realloc(p, sizeof(char *) * (cnt+2));
       p[cnt] = ((struct var_v2 *)VarDB)[i].Name;
